Type handling in its own schema
