+++ schema/src/main/scala/io/shiftleft/codepropertygraph/schema/CpgSchema.scala
@@ -9,17 +9,18 @@ import overflowdb.schema.{Schema, SchemaBuilder}
 class CpgSchema(builder: SchemaBuilder) {
   // the foundation
   val base = Base(builder)
-  val structure = Method(builder, base)
+  val typeDecl = TypeDecl(builder, base)
+  val method = Method(builder, base, typeDecl)
   val metaData = MetaData(builder, base)
-  val enhancements = Enhancements(builder, base, structure)
+  val enhancements = Enhancements(builder, base, method, typeDecl)
 
   // everything else
   val protoSerialize = ProtoSerialize(builder, base)
-  val closure = Closure(builder, base, structure, enhancements)
+  val closure = Closure(builder, base, method, enhancements)
   val finding = Finding(builder, enhancements)
   val operators = Operators(builder)
   val sourceSpecific = Comment(builder, base, enhancements)
-  val tagsAndLocation = TagsAndLocation(builder, base, structure, enhancements)
+  val tagsAndLocation = TagsAndLocation(builder, base, typeDecl, method, enhancements)
 }
 
 object CpgSchema {
